refactor: Relocate `app` to `components` directory
Align with Gutenberg project practices.
